Create Topological sort.cpp
Topological Sort Algorithm | Interview Cake
The topological sort algorithm takes a directed graph and returns an array of the nodes where each node appears before all the nodes it points to. The ordering of the nodes in the array is called a topological ordering. Here's an example: Since node 1 points to nodes 2 and 3, node 1 appears before them in the ordering